LIB-193 modified the datatype of bookId from Long to Integer

+++ src/main/java/com/libraryportal/restapi/dao/CheckoutRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.libraryportal.restapi.entity.Checkout;
 
 public interface CheckoutRepository extends JpaRepository<Checkout, Integer>{
-    Checkout findBookCheckoutByUserEmailAndBookId(String userEmail, Long bookId);
+    Checkout findBookCheckoutByUserEmailAndBookId(String userEmail, Integer bookId);
 }
